+++ tests/DevComputaria.Prompts.Contract.Tests/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace DevComputaria.Prompts.Contract.Tests;
-
-public class Class1
-{
-
-}